Bug found and repaired + new gamma structure

+++ include/typedefs.h
@@ -110,7 +110,7 @@ typedef boost::multi_array<Eigen::MatrixXcd, 10> array_Xcd_d10_eigen;
     std::vector<VdaggerVRandomLookup> rvdaggerv_lookuptable;  
     std::vector<VdaggerVRandomLookup> rvdaggervr_lookuptable;  
   
-    size_t index_of_unity;
+    int index_of_unity;
 
   };
   // ---------------------------------------------------------------------------
